backend/requirements.txt
jieba==0.42.1
faiss-cpu

